+++ packages/create-block/lib/init-block.js
@@ -30,6 +30,7 @@ async function initBlockJSON( {
 	editorStyle,
 	style,
 	render,
+	customBlockJSON,
 } ) {
 	info( '' );
 	info( 'Creating a "block.json" file.' );
@@ -58,6 +59,7 @@ async function initBlockJSON( {
 					editorStyle,
 					style,
 					render,
+					...customBlockJSON,
 				} ).filter( ( [ , value ] ) => !! value )
 			),
 			null,