+++ packages/create-block/lib/scaffold.js
@@ -45,6 +45,8 @@ module.exports = async (
 		style,
 		render,
 		variantVars,
+		customPackageJSON,
+		customBlockJSON,
 	}
 ) => {
 	slug = slug.toLowerCase();
@@ -101,6 +103,8 @@ module.exports = async (
 		editorStyle,
 		style,
 		render,
+		customPackageJSON,
+		customBlockJSON,
 		...variantVars,
 	};