+++ packages/create-block/CHANGELOG.md
@@ -6,6 +6,12 @@
 
 -   Update templates to use the `render` field in `block.json` introduced in WordPress 6.1 ([#44185](https://github.com/WordPress/gutenberg/pull/44185)).
 
+## 4.4.0 (2022-10-19)
+
+### New Feature
+
+-   Add new `customPackageJSON` and `customBlockJSON` keys to allow templates to define custom keys for the resulting `package.json` and `block.json` files respectively.[#44649](https://github.com/WordPress/gutenberg/pull/44649)
+
 ## 4.3.0 (2022-10-05)
 
 ## 4.2.0 (2022-09-21)