+++ packages/server-side-render/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 3.18.0 (2022-10-19)
+
+### Feature
+
+-   Add `skipBlockSupportAttributes` props to prevent duplication of styles in the block wrapper and the `ServerSideRender` components. [#44491](https://github.com/WordPress/gutenberg/pull/44491)
+
 ## 3.17.0 (2022-10-05)
 
 ## 3.16.0 (2022-09-21)